+++ src/matcher/io/io.cpp
@@ -1,5 +1,5 @@
 
-#include "../../include/matcher/io.h"
+#include "../../../include/matcher/io.h"
 #include <hdf5.h>
 #include <iostream>
 #include <string>
