+++ src/matcher/main/matcher_functions.cpp
@@ -1,8 +1,8 @@
-#include "../../include/matcher/matcher_functions.h"
-#include "../../include/matcher/structs_and_classes.h"
-#include "../../include/matcher/read_params.h"
-#include "../../include/matcher/io.h"
-#include "../../include/matcher/utilities.h"
+#include "../../../include/matcher/matcher_functions.h"
+#include "../../../include/matcher/structs_and_classes.h"
+#include "../../../include/matcher/read_params.h"
+#include "../../../include/matcher/io.h"
+#include "../../../include/matcher/utilities.h"
 
 
 //#include <H5Cpp.h>
