+++ src/matcher/utils/structs_and_classes.cpp
@@ -6,11 +6,11 @@
 
 #include <string>
 //#include "../../include/matcher/structs_and_classes.h"
-#include "../../include/matcher/structs_and_classes.h"
+#include "../../../include/matcher/structs_and_classes.h"
 //#include "../../include/matcher/utilities.h"
-#include "../../include/matcher/utilities.h"
+#include "../../../include/matcher/utilities.h"
 //#include "../../include/matcher/matcher_functions.h"
-#include "../../include/matcher/matcher_functions.h"
+#include "../../../include/matcher/matcher_functions.h"
